backend/ai_service.py
The issue is that `emergentintegrations` is a custom library only available within the Emergent platform and cannot be installed on external hosting platforms like Render. Let me fix this by replacing it with the standard Anthropic SDK:
Action: file_editor str_replace /app/backend/ai_service.py --old-str "from emergentintegrations.llm.chat import LlmChat, UserMessage
from dotenv import load_dotenv
import os

load_dotenv()

class AIService:
    def __init__(self):
        self.api_key = os.environ.get(\"EMERGENT_LLM_KEY\")
        self.system_message = \"\"\"You are an AI development assistant for Repbep, a platform that helps developers build applications using AI agents. You provide guidance on:
- Frontend development (React, Tailwind, JavaScript)
- Backend development (FastAPI, Python, MongoDB)
- Code generation and debugging
- Architecture and best practices
- Integration with third-party APIs

Be concise, practical, and provide code examples when helpful. Use markdown formatting for code blocks.\"\"\"
        
        self.chat_instances = {}
    
    async def chat(self, session_id: str, message: str) -> str:
        \"\"\"Send a message to Claude and get a response\"\"\"
        try:
            # Create new chat instance if doesn't exist
            if session_id not in self.chat_instances:
                chat = LlmChat(
                    api_key=self.api_key,
                    session_id=session_id,
                    system_message=self.system_message
                )
                # Configure to use Claude Sonnet 4.5
                chat.with_model(\"anthropic\", \"claude-sonnet-4-5-20250929\")
                self.chat_instances[session_id] = chat
            
            chat = self.chat_instances[session_id]
            
            # Create user message
            user_message = UserMessage(text=message)
            
            # Get response
            response = await chat.send_message(user_message)
            
            return response
            
        except Exception as e:
            print(f\"Error in AI service: {str(e)}\")
            return f\"I apologize, but I encountered an error processing your request. Please try again. Error: {str(e)}\"
    
    def clear_session(self, session_id: str):
        \"\"\"Clear a chat session from memory\"\"\"
        if session_id in self.chat_instances:
            del self.chat_instances[session_id]" --new-str "from anthropic import AsyncAnthropic
from dotenv import load_dotenv
import os

load_dotenv()

class AIService:
    def __init__(self):
        self.api_key = os.environ.get(\"ANTHROPIC_API_KEY\")
        self.client = AsyncAnthropic(api_key=self.api_key)
        self.system_message = \"\"\"You are an AI development assistant for Repbep, a platform that helps developers build applications using AI agents. You provide guidance on:
- Frontend development (React, Tailwind, JavaScript)
- Backend development (FastAPI, Python, MongoDB)
- Code generation and debugging
- Architecture and best practices
- Integration with third-party APIs

Be concise, practical, and provide code examples when helpful. Use markdown formatting for code blocks.\"\"\"
        
        # Store conversation history per session
        self.conversations = {}
    
    async def chat(self, session_id: str, message: str) -> str:
        \"\"\"Send a message to Claude and get a response\"\"\"
        try:
            # Initialize conversation history if doesn't exist
            if session_id not in self.conversations:
                self.conversations[session_id] = []
            
            # Add user message to history
            self.conversations[session_id].append({
                \"role\": \"user\",
                \"content\": message
            })
            
            # Call Claude API
            response = await self.client.messages.create(
                model=\"claude-sonnet-4-20250514\",  # Use Claude Sonnet 4
                max_tokens=4096,
                system=self.system_message,
                messages=self.conversations[session_id]
            )
            
            # Extract response text
            assistant_message = response.content[0].text
            
            # Add assistant response to history
            self.conversations[session_id].append({
                \"role\": \"assistant\",
                \"content\": assistant_message
            })
            
            return assistant_message
            
        except Exception as e:
            print(f\"Error in AI service: {str(e)}\")
            return f\"I apologize, but I encountered an error processing your request. Please try again. Error: {str(e)}\"
    
    def clear_session(self, session_id: str):
        \"\"\"Clear a chat session from memory\"\"\"
        if session_id in self.conversations:
            del self.conversations[session_id]"
Observation: Edit was successful.